move aimakerspace dir under api
