+++ api/vercel.json
@@ -6,4 +6,4 @@
     "routes": [
       { "src": "/(.*)", "dest": "app.py" }
     ]
-  } 
+  }
